Split widget: collections list
